requirements.txt
requests==2.26.0
django-environ==0.4.5
django==3.2.13
djangorestframework==3.12.2
